requirements.txt
dvc[azure]
pyodbc
pandas
pyarrow
python-dotenv


